reorganize bento repository code
